vfs_write: match not only inode number but dev id of superblock
file is uniquely identified by the (inode, dev) pair. So far we catch a write event
for the file with matched inode. But we also need to verify dev id is also matched
to be sure we catch the event for the right file.

Change trigger: many false positives write events on kafka host where kafka log inode
was equal to one of the /etc files inode (we was monitoring /etc/*).

In order to do that, we put in rules table device id as a value (previously dummy value 1
have been written) and on the ebpf side retrieve dev_id from inode superblock.