Fix the "Consumer not found error" (#31)
1. Add similar inode/dev match logic (see ) to the rest of the
probes
2. remove bpf_map_update_elem call from trace_done_path_create probe. This
call creates the situation when we have a rule in ebpf, but don't have a rule on the agent side. In combination with 1 it leads to "Consumer not found error"

Exact scenario:
create subdirectory in directory which has the same inode number as monitored directory but different device id, due to bug 1 we will treat this event as relevant and due to bug 2 we add that inode to our rule map and therefore will monitor any changes for the unrelevant subdirectory. Agent on the contrary doesn't know anything about that directory (because it belongs to diffrent fs) and will spam with "Consumer not found" message on every such event